Extract grid evaluation to util method.
