+++ skfda/_utils/__init__.py
@@ -4,4 +4,5 @@ from ._utils import (_list_of_arrays, _cartesian_product,
                      _check_estimator, parameter_aliases,
                      _to_grid, check_is_univariate,
                      _same_domain, _to_array_maybe_ragged,
-                     _reshape_eval_points)
+                     _reshape_eval_points,
+                     _evaluate_grid)
