+++ skfda/_utils/_utils.py
@@ -127,7 +127,7 @@ def _cartesian_product(axes, flatten=True, return_shape=False):
 
     Examples:
 
-        >>> from skfda.representation._functional_data import _cartesian_product
+        >>> from skfda._utils import _cartesian_product
         >>> axes = [[0,1],[2,3]]
         >>> _cartesian_product(axes)
         array([[0, 2],
@@ -211,6 +211,111 @@ def _reshape_eval_points(eval_points, *, aligned, n_samples, dim_domain):
                              f"evaluation points for each sample.")
 
     return eval_points
+
+
+def _one_grid_to_points(axes, *, dim_domain):
+    """
+    Convert a list of ndarrays, one per domain dimension, in the points.
+
+    Returns also the shape containing the information of how each point
+    is formed.
+    """
+    axes = _list_of_arrays(axes)
+
+    if len(axes) != dim_domain:
+        raise ValueError(f"Length of axes should be "
+                         f"{dim_domain}")
+
+    cartesian, shape = _cartesian_product(axes, return_shape=True)
+
+    # Drop domain size dimension, as it is not needed to reshape the output
+    shape = shape[:-1]
+
+    return cartesian, shape
+
+
+def _evaluate_grid(axes, *, evaluate_method,
+                   n_samples, dim_domain, dim_codomain,
+                   extrapolation=None,
+                   aligned=True):
+    """Evaluate the functional object in the cartesian grid.
+
+    This method is called internally by :meth:`evaluate` when the argument
+    `grid` is True.
+
+    Evaluates the functional object in the grid generated by the cartesian
+    product of the axes. The length of the list of axes should be equal
+    than the domain dimension of the object.
+
+    If the list of axes has lengths :math:`n_1, n_2, ..., n_m`, where
+    :math:`m` is equal than the dimension of the domain, the result of the
+    evaluation in the grid will be a matrix with :math:`m+1` dimensions and
+    shape :math:`n_{samples} x n_1 x n_2 x ... x n_m`.
+
+    If `aligned` is false each sample is evaluated in a
+    different grid, and the list of axes should contain a list of axes for
+    each sample.
+
+    If the domain dimension is 1, the result of the behaviour of the
+    evaluation will be the same than :meth:`evaluate` without the grid
+    option, but with worst performance.
+
+    Args:
+        axes (array_like): List of axes to generated the grid where the
+            object will be evaluated.
+        extrapolation (str or Extrapolation, optional): Controls the
+            extrapolation mode for elements outside the domain range. By
+            default it is used the mode defined during the instance of the
+            object.
+        aligned (bool, optional): If False evaluates each sample
+            in a different grid.
+
+    Returns:
+        (numpy.darray): Numpy array with dim_domain + 1 dimensions with
+            the result of the evaluation.
+
+    Raises:
+        ValueError: If there are a different number of axes than the domain
+            dimension.
+
+    """
+
+    # Compute intersection points and resulting shapes
+    if aligned:
+
+        eval_points, shape = _one_grid_to_points(axes, dim_domain=dim_domain)
+
+    else:
+
+        axes = list(axes)
+
+        if len(axes) != n_samples:
+            raise ValueError("Should be provided a list of axis per "
+                             "sample")
+
+        eval_points, shape = zip(
+            *[_one_grid_to_points(a, dim_domain=dim_domain) for a in axes])
+
+    eval_points = np.array(eval_points)
+
+    # Evaluate the points
+    res = evaluate_method(eval_points,
+                          extrapolation=extrapolation,
+                          aligned=aligned)
+
+    # Reshape the result
+    if aligned:
+
+        res = res.reshape([n_samples] +
+                          list(shape) + [dim_codomain])
+
+    else:
+
+        res = _to_array_maybe_ragged([
+            r.reshape(list(s) + [dim_codomain])
+            for r, s in zip(res, shape)])
+
+    return res
 
 
 def parameter_aliases(**alias_assignments):
