+++ skfda/representation/_functional_data.py
@@ -11,8 +11,7 @@ import pandas.api.extensions
 
 import numpy as np
 
-from .._utils import (_cartesian_product, _list_of_arrays,
-                      _to_array_maybe_ragged, _reshape_eval_points)
+from .._utils import (_evaluate_grid, _reshape_eval_points)
 from .extrapolation import _parse_extrapolation
 
 
@@ -153,107 +152,6 @@ class FData(ABC, pandas.api.extensions.ExtensionArray):
 
         return index
 
-    def _one_grid_to_points(self, axes):
-        """
-        Convert a list of ndarrays, one per domain dimension, in the points.
-
-        Returns also the shape containing the information of how each point
-        is formed.
-        """
-        axes = _list_of_arrays(axes)
-
-        if len(axes) != self.dim_domain:
-            raise ValueError(f"Length of axes should be "
-                             f"{self.dim_domain}")
-
-        cartesian, shape = _cartesian_product(axes, return_shape=True)
-
-        # Drop domain size dimension, as it is not needed to reshape the output
-        shape = shape[:-1]
-
-        return cartesian, shape
-
-    def _evaluate_grid(self, axes, *, extrapolation=None,
-                       aligned=True):
-        """Evaluate the functional object in the cartesian grid.
-
-        This method is called internally by :meth:`evaluate` when the argument
-        `grid` is True.
-
-        Evaluates the functional object in the grid generated by the cartesian
-        product of the axes. The length of the list of axes should be equal
-        than the domain dimension of the object.
-
-        If the list of axes has lengths :math:`n_1, n_2, ..., n_m`, where
-        :math:`m` is equal than the dimension of the domain, the result of the
-        evaluation in the grid will be a matrix with :math:`m+1` dimensions and
-        shape :math:`n_{samples} x n_1 x n_2 x ... x n_m`.
-
-        If `aligned` is false each sample is evaluated in a
-        different grid, and the list of axes should contain a list of axes for
-        each sample.
-
-        If the domain dimension is 1, the result of the behaviour of the
-        evaluation will be the same than :meth:`evaluate` without the grid
-        option, but with worst performance.
-
-        Args:
-            axes (array_like): List of axes to generated the grid where the
-                object will be evaluated.
-            extrapolation (str or Extrapolation, optional): Controls the
-                extrapolation mode for elements outside the domain range. By
-                default it is used the mode defined during the instance of the
-                object.
-            aligned (bool, optional): If False evaluates each sample
-                in a different grid.
-
-        Returns:
-            (numpy.darray): Numpy array with dim_domain + 1 dimensions with
-                the result of the evaluation.
-
-        Raises:
-            ValueError: If there are a different number of axes than the domain
-                dimension.
-
-        """
-
-        # Compute intersection points and resulting shapes
-        if aligned:
-
-            eval_points, shape = self._one_grid_to_points(axes)
-
-        else:
-
-            axes = list(axes)
-
-            if len(axes) != self.n_samples:
-                raise ValueError("Should be provided a list of axis per "
-                                 "sample")
-
-            eval_points, shape = zip(
-                *[self._one_grid_to_points(a) for a in axes])
-
-        eval_points = np.array(eval_points)
-
-        # Evaluate the points
-        res = self.evaluate(eval_points,
-                            extrapolation=extrapolation,
-                            aligned=aligned)
-
-        # Reshape the result
-        if aligned:
-
-            res = res.reshape([self.n_samples] +
-                              list(shape) + [self.dim_codomain])
-
-        else:
-
-            res = _to_array_maybe_ragged([
-                r.reshape(list(s) + [self.dim_codomain])
-                for r, s in zip(res, shape)])
-
-        return res
-
     def _join_evaluation(self, index_matrix, index_ext, index_ev,
                          res_extrapolation, res_evaluation):
         """Join the points evaluated.
@@ -356,9 +254,13 @@ class FData(ABC, pandas.api.extensions.ExtensionArray):
                 aligned=aligned)
 
         if grid:  # Evaluation of a grid performed in auxiliar function
-            return self._evaluate_grid(eval_points,
-                                       extrapolation=extrapolation,
-                                       aligned=aligned)
+            return _evaluate_grid(eval_points,
+                                  evaluate_method=self.evaluate,
+                                  n_samples=self.n_samples,
+                                  dim_domain=self.dim_domain,
+                                  dim_codomain=self.dim_codomain,
+                                  extrapolation=extrapolation,
+                                  aligned=aligned)
 
         if extrapolation is None:
             extrapolation = self.extrapolation
